+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - mapping annotations now correctly extracts the id of the node to apply a new annotation to
 - linking nodes failed to extract node names when graphANNIS responded with a node name only (e. g. in case of "tok" or "node" in a query)
 - linking nodes did not concatenate the values of multiple nodes properly, this is now fixed
+- fixed code of spreadsheet import (merged cells might not have an end column reference)
 
 ## [0.3.1] - 2023-08-04